+++ src/enums/Path.ts
@@ -3,5 +3,6 @@ export enum Path {
     Players = 'players',
     Fight = 'fight',
     Tournaments = 'tournaments',
+    FreeMarket = 'freeMarket',
     Settings = 'settings',
 }