+++ src/modules/reducers.ts
@@ -1,4 +1,5 @@
 export { app } from './app';
+export { freeMarket } from './freeMarket';
 export { players } from './players';
 export { teams } from './teams';
 export { tournaments } from './tournaments';