+++ src/modules/teams/actionTypes.ts
@@ -1,6 +1,7 @@
 export const ADD_TEAM = 'teams/ADD_TEAM';
 export const EDIT_TEAM = 'teams/EDIT_TEAM';
 export const UPDATE_TEAM_ROSTER = 'teams/UPDATE_TEAM_ROSTER';
+export const ASSIGN_PLAYER_TO_ROSTER = 'teams/ASSIGN_PLAYER_TO_ROSTER';
 export const PROMOTE_PLAYER = 'teams/PROMOTE_PLAYER';
 export const REMOVE_TEAM = 'teams/REMOVE_TEAM';
 export const RECORD_TOURNAMENT_RESULT = 'teams/RECORD_TOURNAMENT_RESULT';