+++ src/types/FreeMarketState.ts
@@ -1,0 +1,7 @@
+import { ITeam } from './ITeam';
+import { IPlayer } from './IPlayer';
+
+export interface FreeMarketState {
+    teamsThatDidTransfer: ITeam['id'][];
+    playersWithTeamsActivelyInterested: Record<IPlayer['id'], ITeam['id'][]>;
+}